requirements.txt
line-bot-sdk==3.5.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.0
boto3==1.29.0
pydantic==2.5.1
python-jose==3.3.0